tidy up: impl to .cpp, more const, some copy-and-moves
