+++ include/bronkerbosch.h
@@ -44,46 +44,7 @@ namespace mogli {
     const Product &_product;
 
   public:
-    BronKerbosch(const Product& product, unsigned int min_core_size, unsigned int max_core_size, bool maximum)
-        : _g(product.get_graph())
-        , _product(product)
-        , _n(static_cast<size_t>(lemon::countNodes(_g)))
-        , _cliques()
-        , _bitToNode()
-        , _nodeToBit(_g, std::numeric_limits<size_t>::max())
-        , _bitNeighborhood(_g, BitSet(_n))
-        , _restrictedBitNeighborhood(_g, BitSet(_n))
-        , _min_core_size(min_core_size)
-        , _max_core_size(max_core_size)
-        , _maximum(maximum)
-        , _current_max(0) {
-      // initialize mappings
-      _bitToNode.reserve(_n);
-      size_t i = 0;
-      for (NodeIt v(_g); v != lemon::INVALID; ++v, ++i) {
-        _bitToNode.push_back(v);
-        _nodeToBit[v] = i;
-      }
-
-      // initialize neighborhoods
-      for (NodeIt v(_g); v != lemon::INVALID; ++v, ++i) {
-        BitSet& neighborhood = _bitNeighborhood[v];
-        for (IncEdgeIt e(_g, v); e != lemon::INVALID; ++e) {
-          Node w = _g.oppositeNode(v, e);
-          neighborhood[_nodeToBit[w]] = 1;
-        }
-      }
-
-      // initialize restricted neighborhood mapping
-      for (EdgeIt e(_g); e != lemon::INVALID; ++e) {
-        if (product.is_connectivity_edge(e)) {
-          Node u = _g.u(e);
-          Node v = _g.v(e);
-          _restrictedBitNeighborhood[u][_nodeToBit[v]] = 1;
-          _restrictedBitNeighborhood[v][_nodeToBit[u]] = 1;
-        }
-      }
-    }
+    BronKerbosch(const Product& product, unsigned int min_core_size, unsigned int max_core_size, bool maximum);
 
     void run(int seconds);
 
@@ -99,7 +60,7 @@ namespace mogli {
 
     size_t computeDegeneracy(NodeVector& order);
 
-    void bkPivot(BitSet P, BitSet D, BitSet R, BitSet X, BitSet S,
+    void bkPivot(BitSet P, const BitSet & D, const BitSet & R, BitSet X, const BitSet & S,
                  std::chrono::high_resolution_clock::time_point start, long microseconds);
 
     void report(const BitSet& R);
